% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{anova}
\alias{anova}
\title{applying one factorial anova}
\usage{
anova(data, alpha)
}
\arguments{
\item{data}{dataframe: with cols "value" and "sample"}

\item{alpha}{numeric: the level of significance}
}
\value{
True, if H_0 is believed to be true, Flase if not
}
\description{
one factorial anova tests for differences between samples
}
\examples{
anova(data.frame(sample=factor(c(1,1,2)), value=c(1,2,3)), 0.05)
}
