% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_test.R
\name{complex_test}
\alias{complex_test}
\title{Simulate a complex customisable test}
\usage{
complex_test(test, n, negative)
}
\arguments{
\item{test}{A list of each question of the test.}

\item{n}{The number of repetitions of the test}

\item{negative}{If \code{TRUE} a negative overall score for the
test is allowed, if \code{FALSE} the minimum total score for
 the test is 0. (However individual questions
  can still have negative scores).
 Defaults to \code{FALSE}}
}
\value{
A data frame showing the score achieved for each question
 and the test overall on each run, and corresponding means
 across all runs
}
\description{
Simulates runs of a test with each question in a list. Each question has its
 corresponding parameters from \code{\link{mc}}, \code{\link{mc_all}},
  \code{\link{matching}} and \code{\link{matching_all}}
 in the list entry and takes one of these four forms. An example test is:
 \code{test <- list(q1, q2)} where \code{q1 <- c(type = "mc", options = 5,
  correct = 3, right = 3, wrong = 1, negative = TRUE)} and
 \code{q2 <- c(type = "matching", options = 6,
  right = 2, wrong = 1, negative = FALSE)}
}
\examples{
complex_test(list(c(type = "mc", options = 5, correct = 3, right = 1,
 wrong = 5, negative = TRUE), c(type = "matching", options = 6, right = 4,
  wrong = 1, negative = FALSE)), 1000, FALSE)
complex_test(list(c(type = "mc", options = 5, correct = 3, right = 1,
 wrong = 5, negative = TRUE), c(type = "matching", options = 6, right = 4,
  wrong = 1, negative = FALSE)), 1000, TRUE)
}
