% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_rnorm.R
\name{fixed_rnorm}
\alias{fixed_rnorm}
\title{fixing params on rnorm function}
\usage{
fixed_rnorm(mean, var)
}
\arguments{
\item{mean}{mean of rnorm}

\item{var}{variance of rnorn}
}
\value{
function(n) where n is the sample size of generated sample
}
\description{
fixing params on rnorm function
}
\examples{
fixed_rnorm(10,2)
}
