% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_densities.R
\name{gen_densities}
\alias{gen_densities}
\title{generating densities}
\usage{
gen_densities(k, kind = "norm", offset = 0, application = "to_all", var = 1)
}
\arguments{
\item{k}{number of densities to be created}

\item{kind}{can take on values ["norm", "unif"]:kind of densities to be generated}

\item{offset}{offset of the mean}

\item{application}{can take on values ["to_one", "to_all"] to which the offset applies}

\item{var}{number or vector of length k: variance of the destributions}
}
\value{
list of functions
}
\description{
generates densities for sample generation. Note that this is a function
for general szenarios. I highly encourage experimenting with less general densities.
}
