% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_den.R
\name{gen_rnorm_vector}
\alias{gen_rnorm_vector}
\title{generate vector of rnorm functions}
\usage{
gen_rnorm_vector(params)
}
\arguments{
\item{params}{dataframe of parameters}
}
\value{
vector of rnorm functions of fixed parameters
}
\description{
generate vector of rnorm functions
}
\examples{
gen_rnorm_vector(data.frame(mean=c(1,1,2), var=c(1,1,1)))
}
