% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_wallis.R
\name{kruskal_wallis}
\alias{kruskal_wallis}
\title{applying kruksal wallis to a k-sample}
\usage{
kruskal_wallis(data_raw, alpha)
}
\arguments{
\item{alpha}{value between 0 and 1, level of significance}

\item{sample}{dataframe with cols "value" and "s_cat"}
}
\value{
True if H_0 is believed to be true an Flase if not
}
\description{
applying kruksal wallis to a k-sample
}
\examples{
kruksal_wallis(data.frame(value = c(1,1,3,45,6), sample = factor(c(1,1,2,2,2)))
}
