% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksample.R
\name{ksamples}
\alias{ksamples}
\title{simulating k samples based on given densities}
\usage{
ksamples(k, randfuncs, size = "random")
}
\arguments{
\item{k}{an integer: how many samples should be simulated}

\item{randfuncs}{a vector of functions with length k: the densities for the samples}

\item{size}{a vector of integers with length k: the list of sample-sizes}
}
\value{
a dataframe with columns "value" (realized observation) and "sample" (sample (1 to k) to which the observation belongs to)
}
\description{
ksamples generates independent samples based on given densities.
}
\examples{
ksample(3, c(rnorm, rnorm, function(n) {return(rnorm(n, 0,2))})),  ksample(2, c(rnorm, rnorm))
}
