% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleqs.R
\name{matching}
\alias{matching}
\title{Simulate a customisable matching question}
\usage{
matching(options, right, wrong, negative)
}
\arguments{
\item{options}{The number of pairs that need matching}

\item{right}{The score gained for each correctly matched pair}

\item{wrong}{The score lost for each incorrectly matched pair. Defaults to 0}

\item{negative}{If \code{TRUE} a negative overall score for the
question is allowed, if \code{FALSE} the minimum score
 for the question is 0. Defaults to \code{FALSE}}
}
\value{
The overall mark for the question
}
\description{
Simulate a customisable matching question
}
\examples{
matching(5, 2, 1, negative = TRUE)
matching(8, 1/8)
}
