% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleqs.R
\name{matching_all}
\alias{matching_all}
\title{Simulate a matching question with 1 mark for fully correct, 0 otherwise}
\usage{
matching_all(options)
}
\arguments{
\item{options}{The number of pairs that need matching}
}
\value{
1 if all pairs were correctly matched, 0 otherwise
}
\description{
Simulate a matching question with 1 mark for fully correct, 0 otherwise
}
\examples{
matching_all(6)
matching_all(10)
}
\seealso{
\code{\link{matching}} for more customisable matching questions
}
