% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleqs.R
\name{mc}
\alias{mc}
\title{Simulate a customisable multiple choice question}
\usage{
mc(options, correct, right, wrong, negative)
}
\arguments{
\item{options}{The number of possible options the question has}

\item{correct}{The number of correct answers the question has}

\item{right}{The score gained for each correct answer selected}

\item{wrong}{The score lost for each incorrect answer selected. Defaults to 0}

\item{negative}{If \code{TRUE} a negative overall score for the
question is allowed, if \code{FALSE} the minimum score for
 the question is 0. Defaults to \code{FALSE}}
}
\value{
The overall mark for the question
}
\description{
Simulate a customisable multiple choice question
}
\examples{
mc(5, 3, 2, 1, TRUE)
mc(10, 5, 1/5)
mc(10, 5, 1/5, 1/5)
}
