% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleqs.R
\name{mc_all}
\alias{mc_all}
\title{Simulate a multiple choice question with
 multiple answers, all needed to score}
\usage{
mc_all(options, correct)
}
\arguments{
\item{options}{The number of possible options the question has}

\item{correct}{The number of correct answers the question has}
}
\value{
1 if all correct answers are selected, 0 otherwise
}
\description{
Simulate a multiple choice question with
 multiple answers, all needed to score
}
\examples{
mc_all(5, 3)
mc_all(10, 4)
}
\seealso{
\code{\link{mc}} for more customisable multiple choice questions
}
