% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicmc.R
\name{mctest}
\alias{mctest}
\title{Simulate a basic multiple choice test}
\usage{
mctest(questions, options)
}
\arguments{
\item{questions}{The number of questions in the quiz}

\item{options}{The number of options each question has}
}
\value{
A data frame showing whether each question was
  answered correctly, and the total score achieved
  for the test (out of \code{questions})
}
\description{
Simulate a basic multiple choice test
}
\examples{
mctest(10, 4)
mctest(50, 3)
}
\seealso{
\code{\link{mctest_rep}} for multiple runs of a test

\code{\link{vectortest}} to customise the
 number of options for each question
}
