% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{monte_carlo}
\alias{monte_carlo}
\title{Applying a monte carlo simulation}
\usage{
monte_carlo(eval, n)
}
\arguments{
\item{eval}{stochastic process}

\item{n}{number of executions}
}
\value{
relative and absolute number of occurences
}
\description{
Applying a monte carlo simulation
}
\examples{
monte_carlo(fun, 100)
}
