% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectormc.R
\name{vectortest}
\alias{vectortest}
\title{Simulate a basic multiple choice test with varied option numbers}
\usage{
vectortest(quiz)
}
\arguments{
\item{quiz}{A vector containing the number of options for each question}
}
\value{
A data frame showing whether each question was
  answered correctly, and the total score achieved for the test
}
\description{
Simulate a basic multiple choice test with varied option numbers
}
\examples{
vectortest(c(4,5,6,4,3))
vectortest(rep(4, 10))
}
\seealso{
\code{\link{vectortest_rep}} for multiple runs of such tests
}
