% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectormc.R
\name{vectortest_rep}
\alias{vectortest_rep}
\title{Perform multiple simulations of a multiple
 choice test with varied option numbers}
\usage{
vectortest_rep(quiz, n)
}
\arguments{
\item{quiz}{A vector containing the number of options for each question}

\item{n}{The number of repetitions of the test}
}
\value{
A data frame showing whether each question was
  answered correctly, and the total score achieved for each test,
  along with the mean score across all tests
}
\description{
Perform multiple simulations of a multiple
 choice test with varied option numbers
}
\examples{
vectortest_rep(c(4,5,6,4,3), 1000)
vectortest_rep(rep(4, 10), 30)
}
